code/requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
python-multipart
pydantic~=2.9.2
requests~=2.31.0
sqlalchemy~=2.0.38
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
sumy~=0.11.0
langdetect~=1.0.9
markdown~=3.8
mammoth~=1.9.0
Pillow>=10.0.0
nltk>=3.8.1
passlib[bcrypt]
psycopg2-binary
pgvector
deep-translator
google-cloud-aiplatform
